// WADNR.Web/src/app/shared/generated/model/project-classifications-step-dto.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectClassificationStepItem } from './project-classification-step-item';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectClassificationsStepDto { 
    ProjectID?: number;
    Classifications?: Array<ProjectClassificationStepItem> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectClassificationsStepDtoForm { 
    ProjectID?: FormControl<number>;
    Classifications?: FormControl<Array<ProjectClassificationStepItem>>;
}

export class ProjectClassificationsStepDtoFormControls { 
    public static ProjectID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Classifications = (value: FormControlState<Array<ProjectClassificationStepItem>> | Array<ProjectClassificationStepItem> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectClassificationStepItem>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}