// WADNR.Web/src/app/shared/generated/model/project-classifications-step-request-dto.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectClassificationRequestItem } from './project-classification-request-item';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectClassificationsStepRequestDto { 
    Classifications?: Array<ProjectClassificationRequestItem> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectClassificationsStepRequestDtoForm { 
    Classifications?: FormControl<Array<ProjectClassificationRequestItem>>;
}

export class ProjectClassificationsStepRequestDtoFormControls { 
    public static Classifications = (value: FormControlState<Array<ProjectClassificationRequestItem>> | Array<ProjectClassificationRequestItem> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectClassificationRequestItem>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}