// WADNR.Web/src/app/shared/generated/model/classification-system-with-classifications.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ClassificationOption } from './classification-option';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ClassificationSystemWithClassifications { 
    ClassificationSystemID?: number;
    ClassificationSystemName?: string | null;
    Classifications?: Array<ClassificationOption> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ClassificationSystemWithClassificationsForm { 
    ClassificationSystemID?: FormControl<number>;
    ClassificationSystemName?: FormControl<string>;
    Classifications?: FormControl<Array<ClassificationOption>>;
}

export class ClassificationSystemWithClassificationsFormControls { 
    public static ClassificationSystemID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ClassificationSystemName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Classifications = (value: FormControlState<Array<ClassificationOption>> | Array<ClassificationOption> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ClassificationOption>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}